Add files via upload
I have created three separate classes for agents, the elevator and random data generation. In the Elevator class I have simulated the movements of the elevator as the CheckPermissions method checks if the agent has access to a floor. The agents are managed by threads and when an agent enters the elevator it is saved in ConcurrentQueque.